+++ src/main/kotlin/ru/mipt/npm/gradle/internal/dependencies.kt
@@ -1,4 +1,4 @@
-package ru.mipt.npm.gradle
+package ru.mipt.npm.gradle.internal
 
 import org.gradle.api.Project
 import org.gradle.kotlin.dsl.configure
@@ -6,17 +6,8 @@ import org.gradle.kotlin.dsl.invoke
 import org.jetbrains.kotlin.gradle.dsl.KotlinJsProjectExtension
 import org.jetbrains.kotlin.gradle.dsl.KotlinJvmProjectExtension
 import org.jetbrains.kotlin.gradle.dsl.KotlinMultiplatformExtension
-
-enum class DependencyConfiguration {
-    API,
-    IMPLEMENTATION,
-    COMPILE_ONLY
-}
-
-enum class DependencySourceSet(val setName: String, val suffix: String) {
-    MAIN("main", "Main"),
-    TEST("test", "Test")
-}
+import ru.mipt.npm.gradle.KScienceExtension.DependencyConfiguration
+import ru.mipt.npm.gradle.KScienceExtension.DependencySourceSet
 
 internal fun Project.useDependency(
     vararg pairs: Pair<String, String>,